feat: Added location and tour mockUp model and added context for models

+++ src/Modules/Tours/Explorer.Tours.Core/Domain/Checkpoint.cs
@@ -9,8 +9,8 @@ namespace Explorer.Tours.Core.Domain;
 
     public class Checkpoint : Entity
     {
-        public int TourId { get; init; }
-        public int LocationId { get; init; }
+        public long TourId { get; init; }
+        public long LocationId { get; init; }
         public string Name { get; init; }
         public string Description { get; init; }
         public string ImageUrl { get; init; }
